Quickly hacking 'saving' the sequence back to Volca
Loading the numbered sequence doesn't seem to work - it always returns
the first one, so switch to loading the current sequence (works)
Quickly implement a couple of other Sysex messages including
saving to the current sequence.  This is working, but need
a lot of tidy up.
Next step figure out 'state' for Steps, Notes, etc. and how to
enable us to edit values and have them go back into the sequence
properly when we save.