WIP Zustand store - loading sequence
Create a loose store for the sequences using Zustand and migrate
storing the data into there.
Use in the VolcaFMContext to 'populate' and rejig to lookup using
the sequence number so we can store an array of 'programs'
Consume the store in the Sequence component.

Rough and ready / working
TODO: move the parsing, but how do we 'set' the number in state too?
TODO: add in a function to update a note's data in the sequence in the store